Fixed veh spawn in a multiplayer race with specific config + UX races feedback #91
