+++ BeamJoyInterface/lua/ge/extensions/BJI/ui/WindowRaceSettings/DrawWindowRaceSettings.lua
@@ -242,7 +242,9 @@ end
 
 local function drawHeader(ctxt)
     settings = BJIContext.Scenario.RaceSettings or {}
-    if settings.multi and tlength(BJIContext.Players) < 1 then -- DEBUG 2
+    if settings.multi and tlength(BJIContext.Players) < BJIScenario.get(BJIScenario.TYPES.RACE_MULTI).MINIMUM_PARTICIPANTS then
+        -- when a player leaves then there are not enough players to start
+        BJIToast.warning(BJILang.get("races.preparation.notEnoughPlayers"))
         onClose()
     end
